Update base-system build scripts for LFS 13.0 and refresh extras.
Align cross-toolchain, temp-tools, final-system, and AryaLinux extras with
LFS/BLFS 13.0 package versions and build order; add .gitignore to track only
base-system and applications.

+++ base-system/fix-grub.sh
@@ -8,16 +8,58 @@ set +h
 export MAKEFLAGS="-j `nproc`"
 SOURCE_DIR="/sources"
 LOGFILE="/sources/build-log"
-STEPNAME="091-grub.sh"
-TARBALL="grub-2.06~rc1.tar.xz"
+STEPNAME="fix-grub.sh"
+TARBALL="grub-2.14.tar.xz"
 
 cd $SOURCE_DIR
 
-DIRECTORY=$(tar tf $TARBALL | cut -d/ -f1 | uniq)
+# Installation of pciutils
 
-rm -rf $DIRECTORY
-tar xf $TARBALL
-cd $DIRECTORY
+tar xf pciutils-3.14.0.tar.gz
+cd pciutils-3.14.0
+
+sed -r '/INSTALL/{/PCI_IDS|update-pciids /d; s/update-pciids.8//}' \
+    -i Makefile
+
+make PREFIX=/usr                \
+     SHAREDIR=/usr/share/hwdata \
+     SHARED=yes
+make PREFIX=/usr                \
+     SHAREDIR=/usr/share/hwdata \
+     SHARED=yes                 \
+     install install-lib        &&
+
+chmod -v 755 /usr/lib/libpci.so
+
+cd $SOURCE_DIR
+rm -rf pciutils-3.14.0
+
+# Installation of freetype2
+
+tar xf freetype-2.14.1.tar.xz
+cd freetype-2.14.1
+sed -ri "s:.*(AUX_MODULES.*valid):\1:" modules.cfg &&
+
+sed -r "s:.*(#.*SUBPIXEL_RENDERING) .*:\1:" \
+    -i include/freetype/config/ftoption.h  &&
+
+./configure --prefix=/usr            \
+            --disable-static         \
+            --enable-freetype-config \
+            --with-harfbuzz=no &&
+make
+make install
+
+cd $SOURCE_DIR
+rm -rf freetype-2.14.1
+
+if [ "$TARBALL" != "" ]
+then
+	DIRECTORY=`tar -tf $TARBALL | cut -d/ -f1 | uniq`
+	rm -rf $DIRECTORY
+	tar xf $TARBALL
+	cd $DIRECTORY
+fi
 
 sed -i "s@GNU GRUB  version %s@$OS_NAME $OS_VERSION $OS_CODENAME \- GNU GRUB@g" grub-core/normal/main.c
 
@@ -53,7 +95,7 @@ make
 make install
 
 mkdir -pv /usr/share/fonts/unifont
-gunzip -c ../unifont-7.0.05.pcf.gz > /usr/share/fonts/unifont/unifont.pcf
+zcat ../unifont-17.0.03.pcf.gz > /usr/share/fonts/unifont/unifont.pcf
 grub-mkfont -o /usr/share/grub/unicode.pf2 /usr/share/fonts/unifont/unifont.pcf
 
 cd $SOURCE_DIR
@@ -63,4 +105,4 @@ then
 	rm -rf {gcc,glibc,binutils}-build
 fi
 
-rm -rf $DIRECTORY
+echo "$STEPNAME" | tee -a $LOGFILE
